Add useHydratedStore Move the store hydration and initialisation code into a `useHydratedStore` hook.
