+++ packages/lib-classifier/src/components/Classifier/hooks/index.js
@@ -1,2 +1,3 @@
+export { default as useHydratedStore } from './useHydratedStore'
 export { default as useStore } from './useStore'
 export { default as useWorkflowSnapshot } from './useWorkflowSnapshot'
